requirements.txt
streamlit
tensorflow
keras
opencv-python-headless
Pillow
matplotlib
numpy
gdown
